v3.0.0

+++ easy-es-core/src/main/java/org/dromara/easyes/core/biz/EsIndexParam.java
@@ -38,7 +38,7 @@ public class EsIndexParam {
     /**
      * 索引权重
      */
-    private Float boost;
+    private Double boost;
     /**
      * 查询分词器
      */
